feat: update log output
