+++ apptools/builder.go
@@ -27,7 +27,8 @@ type Builder struct {
 	sourceFactory NewSourceFunc
 	funcs         []NewAppFunc
 
-	enableGops bool
+	enableGops    bool
+	enableMetaKey bool
 }
 
 type App struct {
@@ -99,6 +100,11 @@ func (b *Builder) WithGops() *Builder {
 	return b
 }
 
+func (b *Builder) WithMetaKey() *Builder {
+	b.enableMetaKey = true
+	return b
+}
+
 // Build return cleanup function and error
 func (b *Builder) Build() (*App, func(), error) {
 	app := &App{}
@@ -132,7 +138,10 @@ func (b *Builder) Build() (*App, func(), error) {
 			return nil, nil, err
 		}
 
-		app.logger = WithMetaKeys(logger)
+		app.logger = logger
+		if b.enableMetaKey {
+			app.logger = WithMetaKeys(logger)
+		}
 	}
 
 	var cleanups []func()
