--- schedule/network.go
@@ -1,55 +1,0 @@
-package schedule
-
-import (
-	"fmt"
-	"strings"
-	"time"
-
-	logtos "github.com/hkjojo/go-toolkits/log/v2/kratos"
-	"github.com/shirou/gopsutil/v3/net"
-)
-
-type NetworkMonitor struct {
-	prevNetStats []net.IOCountersStat
-	prevTime     time.Time
-	sent         uint64
-	recv         uint64
-}
-
-func NewNetworkMonitor() *NetworkMonitor {
-	return &NetworkMonitor{}
-}
-
-func (m *NetworkMonitor) collectNetworkStats(log *logtos.ActsHelper) error {
-	currentTime := time.Now()
-	deltaSec := currentTime.Sub(m.prevTime).Seconds()
-	defer func() {
-		m.prevTime = currentTime
-	}()
-
-	var sent, recv uint64
-
-	netStats, _ := net.IOCounters(true)
-	if m.prevNetStats != nil {
-		for i, current := range netStats {
-			if strings.HasPrefix(current.Name, "lo") || strings.HasPrefix(current.Name, "docker") ||
-				strings.HasPrefix(current.Name, "veth") {
-				continue
-			}
-
-			if i < len(m.prevNetStats) {
-				prev := m.prevNetStats[i]
-				sent += current.BytesSent - prev.BytesSent
-				recv += current.BytesRecv - prev.BytesRecv
-			}
-		}
-
-		m.sent = sent
-		m.recv = recv
-		log.Infow(logtos.ModuleSystem, SourceMonitor, fmt.Sprintf("[server] net_sent: %s, %s, net_recv: %s, %s",
-			formatByteSpeed(sent, deltaSec), formatBytes(sent), formatByteSpeed(recv, deltaSec), formatBytes(recv)))
-	}
-	m.prevNetStats = netStats
-
-	return nil
-}
